refactor: lighten up index js
