+++ commands/osCommands.js
